rd_kafka_mock_broker_set_down() now closes the listener instead of closing new connections
This gives a proper "Connection refused" rather than a successful connect
followed by a disconnect.

+++ src/rdkafka_mock_handlers.c
@@ -888,7 +888,7 @@ static int rd_kafka_mock_handle_Metadata(rd_kafka_mock_connection_t *mconn,
                 /* Response: Brokers.Host */
                 rd_kafka_buf_write_str(resp, mrkb->advertised_listener, -1);
                 /* Response: Brokers.Port */
-                rd_kafka_buf_write_i32(resp, mrkb->port);
+                rd_kafka_buf_write_i32(resp, (int32_t)mrkb->port);
                 if (rkbuf->rkbuf_reqhdr.ApiVersion >= 1) {
                         /* Response: Brokers.Rack (Matt's going to love this) */
                         rd_kafka_buf_write_str(resp, mrkb->rack, -1);
@@ -1061,7 +1061,7 @@ rd_kafka_mock_handle_FindCoordinator(rd_kafka_mock_connection_t *mconn,
                 /* Response: NodeId, Host, Port */
                 rd_kafka_buf_write_i32(resp, mrkb->id);
                 rd_kafka_buf_write_str(resp, mrkb->advertised_listener, -1);
-                rd_kafka_buf_write_i32(resp, mrkb->port);
+                rd_kafka_buf_write_i32(resp, (int32_t)mrkb->port);
         }
 
         rd_kafka_mock_connection_send_response(mconn, resp);
